Add files via upload
A python program using the modules: re, pandas, and xlsxwriter to regex a standalone txt file into a dictionary.
From there the dictionary can be  searched as many times as wanted, it can be exported into a .xlsx file, and exited at the user's command.

NOTE: the next iteration of this python program will web scrape for user inputted results from Google to create txt files. The user will be able to search for general results and specify their own location to get proximity results.